#910: Consolidate LengthOf classes
