--- src/main/java/org/cactoos/iterable/LengthOf.java
@@ -1,52 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2017-2018 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.cactoos.iterable;
-
-import org.cactoos.scalar.NumberEnvelope;
-
-/**
- * Length of iterable.
- *
- * <p>There is no thread-safety guarantee.
- *
- * @since 0.1
- */
-public final class LengthOf extends NumberEnvelope {
-
-    /**
-     * Serialization marker.
-     */
-    private static final long serialVersionUID = -7351954368806143451L;
-
-    /**
-     * Ctor.
-     * @param items The array
-     */
-    public LengthOf(final Iterable<?> items) {
-        super(() -> new org.cactoos.iterator.LengthOf(
-            items.iterator()
-        ).doubleValue());
-    }
-
-}
