+++ src/main/java/org/cactoos/scalar/AvgOf.java
@@ -27,7 +27,6 @@ import java.math.BigDecimal;
 import java.math.MathContext;
 import org.cactoos.Scalar;
 import org.cactoos.iterable.IterableOf;
-import org.cactoos.iterable.LengthOf;
 import org.cactoos.iterable.Mapped;
 
 /**
