+++ src/test/java/org/cactoos/func/ChainedTest.java
@@ -26,8 +26,8 @@ package org.cactoos.func;
 import org.cactoos.Func;
 import org.cactoos.iterable.Filtered;
 import org.cactoos.iterable.IterableOf;
-import org.cactoos.iterable.LengthOf;
 import org.cactoos.iterable.Mapped;
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
