+++ src/test/java/org/cactoos/io/StickyTest.java
@@ -26,6 +26,7 @@ package org.cactoos.io;
 import java.net.MalformedURLException;
 import java.net.URL;
 import org.cactoos.func.Repeated;
+import org.cactoos.scalar.LengthOf;
 import org.cactoos.text.TextOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
