+++ src/test/java/org/cactoos/iterable/CycledTest.java
@@ -25,6 +25,7 @@ package org.cactoos.iterable;
 
 import java.util.Collections;
 import org.cactoos.scalar.ItemAt;
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
