+++ src/test/java/org/cactoos/iterable/FilteredTest.java
@@ -23,6 +23,7 @@
  */
 package org.cactoos.iterable;
 
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
