+++ src/test/java/org/cactoos/iterable/StickyTest.java
@@ -26,6 +26,7 @@ package org.cactoos.iterable;
 import java.util.Collections;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.cactoos.list.ListOf;
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
