+++ src/test/java/org/cactoos/iterator/RepeatedTest.java
@@ -23,6 +23,7 @@
  */
 package org.cactoos.iterator;
 
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
