+++ src/test/java/org/cactoos/text/SplitTextTest.java
@@ -23,7 +23,7 @@
  */
 package org.cactoos.text;
 
-import org.cactoos.iterable.LengthOf;
+import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
